rp2040display: try to reduce current on seg/digit drives
v5 hardware runs hot because the NPN transistors
on the segment drives go straight to ground without a
resistor in line. it works, but it's not good for the
hardware long term. drive strength on micropython
appears to be 8 mA by default, which is too generous.
using the minimum drive strength of 2 mA works.

a cvar called low_power_drives can be toggled
off to use default GPIO drive power if need be.